[FIX] web_tour: properly adapt to 9c262fe6f8ca8b47d97d42b4b4b0bcb9229654a7
