+++ addons/web_tour/views/tour_templates.xml
@@ -6,6 +6,7 @@
                 <link rel="stylesheet" type="text/scss" href="/web_tour/static/src/scss/tip.scss"/>
                 <link rel="stylesheet" type="text/scss" href="/web_tour/static/src/scss/keyframes.scss"/>
                 <script type="text/javascript" src="/web_tour/static/src/js/tip.js"></script>
+                <script type="text/javascript" src="/web_tour/static/src/js/tour_utils.js"></script>
                 <script type="text/javascript" src="/web_tour/static/src/js/running_tour_action_helper.js"></script>
                 <script type="text/javascript" src="/web_tour/static/src/js/tour_manager.js"></script>
                 <script type="text/javascript" src="/web_tour/static/src/js/tour_service.js"></script>
